refactor(Pizzeria): major changes

+++ src/main/java/pizzahub/api/repositories/PizzeriaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import pizzahub.api.entities.pizzeria.Pizzeria;
 
-public interface PizzeriaRepository extends JpaRepository <Pizzeria, Short> {
+public interface PizzeriaRepository extends JpaRepository <Pizzeria, Long> {
     Optional<Pizzeria> findByCode (Short code);
 }
